+++ _define.php
@@ -2,7 +2,7 @@
 // +-----------------------------------------------------------------------+
 // | tagDescription - a plugin for dotclear                                |
 // +-----------------------------------------------------------------------+
-// | Copyright(C) 2013-2015 Nicolas Roudaire        http://www.nikrou.net  |
+// | Copyright(C) 2013-2017 Nicolas Roudaire       https://www.nikrou.net  |
 // +-----------------------------------------------------------------------+
 // | This program is free software; you can redistribute it and/or modify  |
 // | it under the terms of the GNU General Public License version 2 as     |
@@ -25,7 +25,7 @@ $this->registerModule(
 	/* Name */				"tagDescription",
 	/* Description*/		"Add description for tags",
 	/* Author */			"Nicolas Roudaire",
-	/* Version */			'0.4.1',
+	/* Version */			'0.5.4',
 	array(
 		'permissions' =>	'contentadmin',
 		'type' => 'plugin',