+++ _public.php
@@ -2,7 +2,7 @@
 // +-----------------------------------------------------------------------+
 // | tagDescription - a plugin for dotclear                                |
 // +-----------------------------------------------------------------------+
-// | Copyright(C) 2013-2015 Nicolas Roudaire        http://www.nikrou.net  |
+// | Copyright(C) 2013-2017 Nicolas Roudaire       https://www.nikrou.net  |
 // +-----------------------------------------------------------------------+
 // | This program is free software; you can redistribute it and/or modify  |
 // | it under the terms of the GNU General Public License version 2 as     |
@@ -22,6 +22,7 @@
 if (!defined('DC_RC_PATH')) { return; }
 
 $core->tpl->addValue('TagDescription', array('tagDescriptionTpl', 'TagDescription'));
+$core->tpl->addValue('TagTitle', array('tagDescriptionTpl', 'TagTitle'));
 
 if ($core->url->type == 'default') {
     $_ctx->tag_manager = new tagManager($core);